portfolio/05-semantic-search/requirements.txt
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
fastapi>=0.100.0
uvicorn>=0.23.0
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
